+++ src/narf/math/floats.cpp
@@ -1,9 +1,10 @@
+#include <tgmath.h>
 #include "narf/math/floats.h"
 #include <cmath>
 #include <float.h>
 
 bool narf::math::AlmostEqualDiff(float A, float B, float maxDiff) {
-	float absDiff = fabs(A - B);
+	float absDiff = (float)fabs(A - B);
 	return absDiff <= maxDiff;
 }
 
@@ -20,7 +21,7 @@ bool narf::math::AlmostEqualUlps(float A, float B, int maxUlpsDiff) {
 bool narf::math::AlmostEqualUlpsAndAbs(float A, float B, float maxDiff, int maxUlpsDiff) {
 	// Check if the numbers are really close -- needed
 	// when comparing numbers near zero.
-	float absDiff = fabs(A - B);
+	float absDiff = (float)fabs(A - B);
 	if (absDiff <= maxDiff)
 		return true;
 
@@ -32,7 +33,7 @@ bool narf::math::AlmostEqualUlpsAndAbs(float A, float B, float maxDiff, int maxU
 		return false;
 
 	// Find the difference in ULPs.
-	int ulpsDiff = std::abs(uA.i - uB.i);
+	int ulpsDiff = (int)std::abs(uA.i - uB.i);
 	if (ulpsDiff <= maxUlpsDiff)
 		return true;
 
@@ -42,12 +43,12 @@ bool narf::math::AlmostEqualUlpsAndAbs(float A, float B, float maxDiff, int maxU
 bool narf::math::AlmostEqualRelativeAndAbs(float A, float B, float maxDiff, float maxRelDiff) {
 	// Check if the numbers are really close -- needed
 	// when comparing numbers near zero.
-	float diff = fabs(A - B);
+	float diff = (float)fabs(A - B);
 	if (diff <= maxDiff)
 		return true;
 
-	A = fabs(A);
-	B = fabs(B);
+	A = (float)fabs(A);
+	B = (float)fabs(B);
 	float largest = (B > A) ? B : A;
 
 	if (diff <= largest * maxRelDiff)
